move logic to store synced translations to the SyncRepository
